+++ fuel-indexer/Cargo.toml
@@ -13,7 +13,6 @@ async-trait = "0.1"
 axum = { version = "0.4", features = ["multipart"] }
 bincode = "1.3.3"
 cfg-if = "1.0"
-clap = { version = "3.1", features = ["cargo", "derive", "env"] }
 fuel-core = { version = "0.10", default-features = false }
 fuel-gql-client = { version = "0.10" }
 fuel-indexer-database = { path =  "../fuel-indexer-database" }
