+++ fuel-indexer/src/service.rs
@@ -1,8 +1,5 @@
 use crate::{
-    config::{FuelNodeConfig, IndexerConfig},
-    manifest::Module,
-    Executor, IndexerResult, Manifest, NativeIndexExecutor, SchemaManager,
-    WasmIndexExecutor,
+    Executor, IndexerResult, NativeIndexExecutor, SchemaManager, WasmIndexExecutor,
 };
 use async_std::{fs::File, io::ReadExt, sync::Arc};
 use fuel_gql_client::client::{
@@ -10,6 +7,8 @@ use fuel_gql_client::client::{
 };
 use fuel_indexer_database::{queries, IndexerConnectionPool};
 use fuel_indexer_database_types::IndexAssetType;
+use fuel_indexer_lib::config::{FuelNodeConfig, IndexerConfig};
+use fuel_indexer_lib::manifest::{Manifest, Module};
 use fuel_indexer_lib::utils::AssetReloadRequest;
 use fuel_indexer_schema::{Address, BlockData, Bytes32};
 use futures::stream::{futures_unordered::FuturesUnordered, StreamExt};
